@@ -4,7 +4,7 @@ version: 2.1.1
 homepage: https://github.com/ShekarMudaliyar/social_share
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
