@@ -8,12 +8,12 @@ import 'package:path_provider/path_provider.dart';
 class SocialShare {
   static const MethodChannel _channel = const MethodChannel('social_share');
 
-  static Future<String> shareInstagramStory(
+  static Future<String?> shareInstagramStory(
       String imagePath, {
-        String backgroundTopColor,
-        String backgroundBottomColor,
-        String attributionURL,
-        String backgroundImagePath,
+        String? backgroundTopColor,
+        String? backgroundBottomColor,
+        String? attributionURL,
+        String? backgroundImagePath,
       }) async {
     Map<String, dynamic> args;
     if (Platform.isIOS) {
@@ -45,7 +45,7 @@ class SocialShare {
       file = await File(stickerAssetPath).create();
       file.writeAsBytesSync(stickerAssetAsList);
 
-      String backgroundAssetName;
+      String? backgroundAssetName;
       if (backgroundImagePath != null) {
         File backgroundImage = File(backgroundImagePath);
         Uint8List backgroundImageData = backgroundImage.readAsBytesSync();
@@ -64,19 +64,19 @@ class SocialShare {
         "attributionURL": attributionURL,
       };
     }
-    final String response = await _channel.invokeMethod(
+    final String? response = await _channel.invokeMethod(
       'shareInstagramStory',
       args,
     );
     return response;
   }
 
-  static Future<String> shareFacebookStory(
+  static Future<String?> shareFacebookStory(
       String imagePath,
       String backgroundTopColor,
       String backgroundBottomColor,
       String attributionURL,
-      {String appId}) async {
+      {String? appId}) async {
     Map<String, dynamic> args;
     if (Platform.isIOS) {
       args = <String, dynamic>{
@@ -103,19 +103,19 @@ class SocialShare {
         "appId": appId
       };
     }
-    final String response =
+    final String? response =
     await _channel.invokeMethod('shareFacebookStory', args);
     return response;
   }
 
-  static Future<String> shareTwitter(String captionText,
-      {List<String> hashtags, String url, String trailingText}) async {
+  static Future<String?> shareTwitter(String captionText,
+      {List<String>? hashtags, String? url, String? trailingText}) async {
     Map<String, dynamic> args;
     String modifiedUrl;
     if (Platform.isAndroid) {
-      modifiedUrl = Uri.parse(url).toString().replaceAll('#', "%23");
+      modifiedUrl = Uri.parse(url!).toString().replaceAll('#', "%23");
     } else {
-      modifiedUrl = Uri.parse(url).toString();
+      modifiedUrl = Uri.parse(url!).toString();
     }
     if (hashtags != null && hashtags.isNotEmpty) {
       String tags = "";
@@ -136,13 +136,13 @@ class SocialShare {
         (trailingText == null || trailingText.isEmpty) ? "" : trailingText
       };
     }
-    final String version = await _channel.invokeMethod('shareTwitter', args);
+    final String? version = await _channel.invokeMethod('shareTwitter', args);
     return version;
   }
 
-  static Future<String> shareSms(String message,
-      {String url, String trailingText}) async {
-    Map<String, dynamic> args;
+  static Future<String?> shareSms(String message,
+      {String? url, String? trailingText}) async {
+    Map<String, dynamic>? args;
     if (Platform.isIOS) {
       if (url == null) {
         args = <String, dynamic>{
@@ -157,23 +157,23 @@ class SocialShare {
       }
     } else if (Platform.isAndroid) {
       args = <String, dynamic>{
-        "message": message + url + trailingText,
+        "message": message + url! + trailingText!,
       };
     }
-    final String version = await _channel.invokeMethod('shareSms', args);
+    final String? version = await _channel.invokeMethod('shareSms', args);
     return version;
   }
 
-  static Future<bool> copyToClipboard(content) async {
+  static Future<bool?> copyToClipboard(content) async {
     final Map<String, String> args = <String, String>{
       "content": content.toString()
     };
-    final bool response = await _channel.invokeMethod('copyToClipboard', args);
+    final bool? response = await _channel.invokeMethod('copyToClipboard', args);
     return response;
   }
 
-  static Future<bool> shareOptions(String contentText,
-      {String imagePath}) async {
+  static Future<bool?> shareOptions(String contentText,
+      {String? imagePath}) async {
     Map<String, dynamic> args;
     if (Platform.isIOS) {
       args = <String, dynamic>{"image": imagePath, "content": contentText};
@@ -193,24 +193,24 @@ class SocialShare {
         args = <String, dynamic>{"image": imagePath, "content": contentText};
       }
     }
-    final bool version = await _channel.invokeMethod('shareOptions', args);
+    final bool? version = await _channel.invokeMethod('shareOptions', args);
     return version;
   }
 
-  static Future<String> shareWhatsapp(String content) async {
+  static Future<String?> shareWhatsapp(String content) async {
     final Map<String, dynamic> args = <String, dynamic>{"content": content};
-    final String version = await _channel.invokeMethod('shareWhatsapp', args);
+    final String? version = await _channel.invokeMethod('shareWhatsapp', args);
     return version;
   }
 
-  static Future<Map> checkInstalledAppsForShare() async {
-    final Map apps = await _channel.invokeMethod('checkInstalledApps');
+  static Future<Map?> checkInstalledAppsForShare() async {
+    final Map? apps = await _channel.invokeMethod('checkInstalledApps');
     return apps;
   }
 
-  static Future<String> shareTelegram(String content) async {
+  static Future<String?> shareTelegram(String content) async {
     final Map<String, dynamic> args = <String, dynamic>{"content": content};
-    final String version = await _channel.invokeMethod('shareTelegram', args);
+    final String? version = await _channel.invokeMethod('shareTelegram', args);
     return version;
   }
 
